requirements.txt
fastapi==0.115.12
pandas==2.2.3
psycopg2_binary==2.9.10
pydantic==2.11.4
python-dotenv==1.1.0
Requests==2.32.3
streamlit==1.45.0
transformers==4.51.3
uvicorn==0.35.0
sqlparse==0.5.3
torch==2.7.1
accelerate==1.9.0
optimum==1.26.1
auto-gptq==0.7.1